moved contents of test_job_manager_test_env to test_job_manager
